Fixes #40 (CouchDB auth) and begins work on #39 (bulkloader for CellIndex).
For issue #40, we authenticate to couch using cookies via the `/session` endpoint:
http://guide.couchdb.org/draft/security.html

Credentials are expected in a `/app/couchdb-creds.txt` file of the form
`user:password` with a newline character at the end. These credentials are
authenticated against the database reader list on iris. Note that the
`/app/couchdb-creds.txt` file is in `.gitignore` and shouldn't be pushed to
the repo.

For issue #39, the bulkloader for `CellIndex` requires a transformation of the
CSV files coming out of Starspan. The transformation creates a new CSV file with
header `cellkey,varvals` where `cellkey` is the cell key and `varvals` is a JSON
representation of a dictionary containing all variable name/value pairs. So the
new CSV file contains one cell per line. This needs to be rolled into the `sdl.py`
script.